+++ api/ciclo_ruta.py
@@ -16,6 +16,7 @@ def getciclo_ruta():
 
 @ruta_ciclo_ruta.route("/saveciclo_ruta", methods=["POST"])
 def saveciclo_ruta():
+ try:
     id_ciclo_ruta = request.json['id_ciclo_ruta']
     id_ciclo = request.json['id_ciclo']
     id_ruta = request.json['id_ruta']
@@ -23,10 +24,15 @@ def saveciclo_ruta():
     db.session.add(new_ciclo_ruta)
     db.session.commit()
     return "Datos guardados con exitos"
+ except Exception as e:
+        return f"Hubo un error {str(e)}"
 
-@ruta_ciclo_ruta.route("/updateciclo_ruta", methods=["PUT"])
-def updateciclo_ruta():
-    id_ciclo_ruta = request.json['id_ciclo_ruta']
+@ruta_ciclo_ruta.route("/updateciclo_ruta/<id>", methods=["PUT"])
+def updateciclo_ruta(id):
+  try:
+    id_ciclo_ruta =  ciclo_ruta.query.get(id)
+    if not id_ciclo_ruta:
+       return " ciclo_ruta no esta registrado"
     id_ciclo = request.json['id_ciclo']
     id_ruta = request.json['id_ruta']
     nciclo_ruta = ciclo_ruta.query.get(id_ciclo_ruta,id_ciclo, id_ruta) #Select * from Cliente where id = id
@@ -35,12 +41,21 @@ def updateciclo_ruta():
     nciclo_ruta.id_ruta=id_ruta
     db.session.commit()
     return "Datos Actualizado con exitos"
+  except Exception as e:
+        return f"Hubo un error {str(e)}"
 
 @ruta_ciclo_ruta.route("/deleteciclo_ruta/<id>", methods=["DELETE"])
 def deleteciclo_ruta(id):
-    id_ciclo_ruta = request.json['id_ciclo_ruta']
+ try:
+    id_ciclo_ruta = ciclo_ruta.query.get(id)
+    if not id_ciclo_ruta:
+       return " ciclo_ruta no esta registrado"
     ciclo_rutax = ciclo_ruta.query.get(id_ciclo_ruta)
     db.session.delete(ciclo_rutax)
     db.session.commit()
+    return "se ha eliminado"
+ except Exception as e:
+        return f"Hubo un error{str(e)}"
+
